requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pinecone==5.0.0
google-generativeai==0.3.2
langchain==0.0.350
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0
faiss-cpu==1.7.4
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
